5218: Separate connector and proxy nodes
Deploy EidasNode twice, once to behave as a connector, once to behave as
a Proxy. We've had to fiddle around with some environment variables to
achieve this, seems to work locally now.

The big gotcha is the serviceMetadataGeneratorIDP settings, which set
how the Node describes itself in its metadata. We've set *both* nodes
(since there's only one source of config) to describe themselves as
proxy nodes. This is a bit confusing, but it seems to work for the
journey we're testing.

It's also a bit weird that we need to set the specificidpredirect.url
setting - surely the IdP should know where to redirect the user back to
from the AuthnRequest? Happy to ignore this weirdness for now.

Authors: @richardtowers @niyioyelade